Don't use './setup.py install'.

+++ install_test_dependencies.sh
@@ -4,7 +4,10 @@
 
 set -e
 
-sudo apt-get install -y python-setuptools
-sudo pip install pytest
+if ! dpkg -l python-setuptools > /dev/null 2>&1; then
+	sudo apt-get install -y python-setuptools
+fi
+
 sudo pip install coverage
+sudo pip install pytest
 sudo pip install pytest_cov
